Update the readme to be accurate
I have updated the readme to match the accurate path for the configuration.  There is not a branch for 1.1.1 so I was unable to try to make the PR against that version.  I have not tested this for v4 so this PR may be irrelevant for that version, but the README.md on version 1.1.1 is incorrect.